chore: move span method for context in trace API #40
